fix(0x02): update `wrapper` function
This change update the return type annotation of the wrapper function.
It also adds type checking for the `self._redis` instance before using
it.
Another change is to the return value of the wrapper function. We now
return the return value of the method callable after passing it all the
args and kwargs. This ensures that the method it is used on works in all
situations and preserves its initial behaviour.